requirements.txt
fastapi
httpx
uvicorn
python-dotenv
google-generativeai
fuzzywuzzy
numpy
pydantic